todo placeholders and forloop path initiation
